Updated makefile, directory layout, fixed load/unload script, and scull works.
makefile was updated to move *.o to objects, ko to modules, and to delete
all appropriate files including symvers, order and so on.

Unload and load script moved out of src and into head directory. Modified
to handle directory layout and tested.

scull device /dev/scull was tested and accepted input, not test on reading
output back out.